requirements.txt
python_gitlab==2.7.1

